# quodom-web/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quodom API
  version: 1.0.0
  description: |
    Contrato entre la app y el API. **Este archivo manda**: todo cambio de API
    empieza acá y después va el código. El servidor valida los pedidos contra
    este archivo, los tests validan las respuestas, y la app genera sus tipos
    con `npm run api-types`.
servers:
  - url: /
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: quodom_session
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Pedido inválido o regla de negocio rota
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Sin sesión o sesión inválida
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Conflict:
      description: Choca con un Quodom abierto (rubro_duplicado / rubro_mismatch)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  parameters:
    IntId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    QuodomId:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    Error:
      type: object
      required: [res]
      properties:
        res: { type: boolean, enum: [false] }
        message: { type: string }
        error: { type: string }
        exp: { type: boolean }
      additionalProperties: true
...
    OkMessage:
      type: object
      required: [res, message]
      properties:
        res: { type: boolean }
        message: { type: string }
...
    Atributo:
      type: object
      required: [valoratributo]
      properties:
        valoratributo: { type: string }
        orden: { type: integer, nullable: true }
        esvendedor: { type: string, nullable: true }
...
    VQuodom:
      type: object
      required: [id, descripcion, estado, nro, createdBy]
      properties:
        id: { type: string }
        descripcion: { type: string }
        createdBy: { type: string }
        estado: { type: string, enum: [CREADO, ENVIADO] }
        idrubro: { type: integer, nullable: true }
        nombrerubro: { type: string, nullable: true }
        porccompletado: { type: number, nullable: true }
        cantproductos: { type: integer, nullable: true }
        fechavencimientoenvio: { type: string, format: date-time, nullable: true }
        diasparavencimientoenvio: { type: integer, nullable: true }
        diasparavencimientoaceptacion: { type: integer, nullable: true }
        iddireccion: { type: string, nullable: true }
        nro: { type: string }
        fechaenvio: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    VQuodomLine:
      type: object
      required: [id, idquodom, idproducto, cantidad]
      properties:
        id: { type: integer }
        idquodom: { type: string }
        idproducto: { type: integer }
        cantidad: { type: number }
        categoria: { type: integer, nullable: true }
        categoriaPadre: { type: integer, nullable: true }
        nombreCategoria: { type: string, nullable: true }
        nombreProducto: { type: string, nullable: true }
        detalleProducto: { type: string, nullable: true }
        marca: { type: string, nullable: true }
        unidad: { type: string, nullable: true }
        atributosFaltantes: { type: integer, nullable: true }
        imagen: { type: string, nullable: true }
        refreshImagen: { type: string, nullable: true }
        atributo1: { type: string, nullable: true, description: 'Lo que eligió el usuario' }
        atributo2: { type: string, nullable: true }
        nombreAtributo1: { type: string, nullable: true, description: 'Nombre del grupo' }
        nombreAtributo2: { type: string, nullable: true }
    Direccion:
      type: object
      required: [id, userid]
      properties:
        id: { type: integer }
        userid: { type: string }
        provincia: { type: string, nullable: true }
        partido: { type: string, nullable: true }
        localidad: { type: string, nullable: true }
        direccion: { type: string, nullable: true }
        calle: { type: string, nullable: true }
        numero: { type: string, nullable: true }
        piso: { type: string, nullable: true }
        cp: { type: string, nullable: true }
        alias: { type: string, nullable: true }
        default: { anyOf: [{ type: boolean, nullable: true }, { type: integer, nullable: true }] }
        idprovincia: { type: string, nullable: true }
        idpartido: { type: string, nullable: true }
        observaciones: { type: string, nullable: true }
    DireccionInput:
      type: object
      properties:
        alias: { type: string }
        calle: { type: string }
        numero: { type: string }
        piso: { type: string }
        cp: { type: string }
        localidad: { type: string }
        direccion: { type: string }
        observaciones: { type: string }
        idprovincia: { type: integer }
        default: { type: boolean }
...
paths:
...
  /users/direcciondefault:
    get:
      tags: [users]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: La dirección por defecto, o la primera, o null
          content: { application/json: { schema: { oneOf: [{ $ref: '#/components/schemas/Direccion' }, { type: string, nullable: true, enum: [null] }] } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /users/{id}:
    get:
      tags: [users]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200':
          description: Sólo el propio usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  username: { type: string }
                  email: { type: string }
                  nombre: { type: string, nullable: true }
                  apellido: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [users]
      summary: Borrar un usuario (sólo admin)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /quodom/getQuodomCreados:
    get:
      tags: [quodom]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [res, data]
                properties:
                  res: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/VQuodom' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /quodom/misQuodom:
    get:
      tags: [quodom]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/VQuodom' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /quodom/porccompletado/{id}:
    get:
      tags: [quodom]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200':
          description: OK (null si no existe)
          content:
            application/json:
              schema:
                type: object
                nullable: true
                properties:
                  cantproductos: { type: integer, nullable: true }
                  porccompletado: { type: number, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /quodom/whatsapp/{id}:
    get:
      tags: [quodom]
      summary: Arma el link wa.me y pasa el Quodom a ENVIADO
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [res, link]
                properties:
                  res: { type: boolean }
                  link: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /quodom/activo/{idrubro}:
    get:
      tags: [quodom]
      summary: El Quodom abierto (CREADO) del usuario en ese rubro, o null
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: idrubro, in: path, required: true, schema: { type: string }, description: 'Entero positivo; el handler lo valida y contesta idrubro_invalido' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [res]
                properties:
                  res: { type: boolean }
                  data: { oneOf: [{ $ref: '#/components/schemas/VQuodom' }, { type: string, nullable: true, enum: [null] }] }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /quodom/create:
    post:
      tags: [quodom]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [descripcion, idrubro]
              properties:
                descripcion: { type: string }
                idrubro: { type: integer }
                iddireccion: { type: integer, nullable: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [res, idquodom]
                properties:
                  res: { type: boolean }
                  idquodom: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /quodom/repetir/{id}:
    post:
      tags: [quodom]
      summary: Crea un Quodom nuevo con las líneas de uno ENVIADO
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [res, idquodom]
                properties:
                  res: { type: boolean }
                  idquodom: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /quodom/{id}:
    get:
      tags: [quodom]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/VQuodom' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [quodom]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                descripcion: { type: string }
                iddireccion: { type: integer, nullable: true }
      responses:
        '200': { description: 'OK (el cuerpo es true)', content: { application/json: { schema: { type: boolean } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [quodom]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /quodom_lines/lines/{id}:
    get:
      tags: [quodom_lines]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IntId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  idproducto: { type: integer }
                  detalleProducto: { type: string, nullable: true }
                  cantidad: { type: number }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /quodom_lines/atributos:
    get:
      tags: [quodom_lines]
      summary: Valores elegibles de un grupo de atributo de un producto
      parameters:
        - { name: idproducto, in: query, required: false, schema: { type: integer } }
        - { name: nombreatributo, in: query, required: false, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Atributo' } } } } }
  /quodom_lines/add:
    post:
      tags: [quodom_lines]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idquodom, idproducto, cantidad]
              properties:
                idquodom: { type: string }
                idproducto: { type: integer }
                cantidad: { type: integer }
                nombreProducto: { type: string }
                atributo1: { type: string }
                atributo2: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [res, id]
                properties:
                  res: { type: boolean }
                  id: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /quodom_lines/{id}:
    get:
      tags: [quodom_lines]
      summary: Líneas de un Quodom (id = id del Quodom)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/VQuodomLine' } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [quodom_lines]
      summary: Cambiar cantidad o atributo elegido (id = id de la línea)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cantidad: { type: integer }
                atributo1: { type: string }
                atributo2: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [quodom_lines]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /user_direcciones/direcciondefault:
    get:
      tags: [user_direcciones]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200': { description: 'OK (null si no tiene)', content: { application/json: { schema: { oneOf: [{ $ref: '#/components/schemas/Direccion' }, { type: string, nullable: true, enum: [null] }] } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /user_direcciones/create:
    post:
      tags: [user_direcciones]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DireccionInput' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /user_direcciones/prin/{id}:
    put:
      tags: [user_direcciones]
      summary: Marca la dirección como principal
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IntId' }]
      responses:
        '200': { description: 'OK; el cuerpo lo devuelve el controller tal cual (ver Task 6)', content: { application/json: { schema: {} } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /user_direcciones/{id}:
    get:
      tags: [user_direcciones]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IntId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Direccion' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [user_direcciones]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IntId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DireccionInput' } } }
      responses:
        '200': { description: 'OK; el cuerpo lo devuelve el controller tal cual (ver Task 6)', content: { application/json: { schema: {} } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [user_direcciones]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IntId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...